requirements.txt
streamlit
streamlit-option-menu
pillow
google-generativeai